dicotomiaExamen

+++ FUNCTIONS/f3.f95
@@ -1,8 +1,8 @@
 function f3(x) result(res)
 
     implicit none
-    real(4), intent(in) :: x
-    real(4) :: res
+    real(8), intent(in) :: x
+    real(8) :: res
 
     res = 0.73169 - x + 0.09341*sin(x)
